[keyframe] Add first preliminary frame selection based on OF on borders
Using the computed optical flow on each border, drop frames that are
do not contain enough flow. The threshold used to determine whether a
border contains enough flow is the median motion of that same border.

If at least one border's flow is below its median, the frame is dropped.