[keyframe] Add rough refinement after the initial frame selection
Once frames have been selected based on their optical flow scores on
the borders, if the "refine" option is on, perform a rough refinement.

The frames selected with their motion scores are first clusterized.
Then, for each cluster, the sharpest frame is selected. This is a flawed
method as some clusters contain lots of frames, and only selecting
a single frame implies losing plenty of meaningful frames.

The method will later be extended to select more than 1 frame per cluster,
depending on the cluster's size, the provided parameters and the relevance
of the frames within a cluster.