[keyframe] Add smart keyframe selection based on optical flow by cell
Determine subsequences within the video sequence based on the accumulated
pixel displacement, obtained with the optical flow by cell approach.

Once subsequences are set (with respect to constraints like the minimum
and maximum number of keyframes), a keyframe is selected within that
subsequence, based on both its sharpness score and its temporal position
within the subsequence (frames in the middle of the subsequence have better
weights than those at the very beginning or very end of the subsequence).
The goal is to maximize sharpness while still preserving temporal
representation: if the last frame of a subsequence and the first frame of
the next subsequence are both the sharpest of their respective
subsequences, we do not want to select them both as they are consecutive
and probably won't bring different information.